extract filter logic
